move events dtos to common module
