+++ common/src/main/java/io/apicurio/registry/events/dto/ArtifactId.java
