+++ common/src/main/java/io/apicurio/registry/events/dto/ArtifactRuleChange.java
