+++ common/src/main/java/io/apicurio/registry/events/dto/ArtifactStateChange.java
