+++ common/src/main/java/io/apicurio/registry/events/dto/RegistryEventType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.apicurio.registry.events;
+package io.apicurio.registry.events.dto;
 
 /**
  * @author Fabian Martinez
